Move manual endpoint impl test into its own file
